# Asist.swagger.json
{"info":{"title":"My Project","description":"","version":"1.0.0"},"tags":[{"name":"Asist.io"},{"name":"Asist.io/Usuarios"}],"paths":{"/usuario/registro":{"post":{"summary":"Registro","deprecated":false,"description":"","tags":["Asist.io/Usuarios"],"parameters":[{"name":"body","in":"body","schema":{"type":"object","properties":{},"x-apidog-ignore-properties":[],"x-apidog-orders":[]}}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{},"x-apidog-ignore-properties":[],"x-apidog-orders":[]}}},"security":[],"x-apidog-folder":"Asist.io/Usuarios","x-apidog-status":"developing","x-run-in-apidog":"https://app.apidog.com/web/project/412546/apis/api-5757482-run","consumes":["application/json"],"produces":["application/json"]}},"/auth/login":{"post":{"summary":"Login","deprecated":false,"description":"","tags":["Asist.io/Usuarios"],"parameters":[{"name":"body","in":"body","schema":{"type":"object","properties":{},"x-apidog-ignore-properties":[],"x-apidog-orders":[]}}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{},"x-apidog-ignore-properties":[],"x-apidog-orders":[]}}},"security":[],"x-apidog-folder":"Asist.io/Usuarios","x-apidog-status":"developing","x-run-in-apidog":"https://app.apidog.com/web/project/412546/apis/api-5757948-run","consumes":["application/json"],"produces":["application/json"]}},"/usuario/reenviar-correo-confirmacion":{"post":{"summary":"Reenviar Correo de Confirmacion","deprecated":false,"description":"","tags":["Asist.io/Usuarios"],"parameters":[{"name":"correo","in":"query","description":"","required":false,"type":"string","x-example":"enzo.meneghini@hotmail.com"}],"responses":{"200":{"description":"OK","schema":{"type":"object","properties":{"Mensaje":{"type":"string"}},"required":["Mensaje"],"x-apidog-ignore-properties":[],"x-apidog-orders":["Mensaje"]}}},"security":[],"x-apidog-folder":"Asist.io/Usuarios","x-apidog-status":"developing","x-run-in-apidog":"https://app.apidog.com/web/project/412546/apis/api-5849434-run","produces":["application/json"]}},"/usuario/cambiar-contrasena/b243eb57-7a01-408e-8996-d37ba7183d45":{"patch":{"summary":"Cambiar Contraseña (no logged)","deprecated":false,"description":"","tags":["Asist.io/Usuarios"],"parameters":[{"name":"contrasena","in":"query","description":"","required":false,"type":"string","x-example":"contraseña.2"}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{},"x-apidog-ignore-properties":[],"x-apidog-orders":[]}}},"security":[],"x-apidog-folder":"Asist.io/Usuarios","x-apidog-status":"developing","x-run-in-apidog":"https://app.apidog.com/web/project/412546/apis/api-5849465-run","produces":["application/json"]}},"/usuario/cambiar-contrasena":{"patch":{"summary":"Cambiar Contraseña (logged)","deprecated":false,"description":"","tags":["Asist.io/Usuarios"],"parameters":[{"name":"contrasenaNueva","in":"query","description":"","required":false,"type":"string","x-example":"contrasena.1"},{"name":"contrasenaActual","in":"query","description":"","required":false,"type":"string","x-example":"meneghini.01"},{"name":"correo","in":"query","description":"","required":false,"type":"string","x-example":"enzo.meneghini@hotmail.com"}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{},"x-apidog-ignore-properties":[],"x-apidog-orders":[]}}},"security":[{"bearer":[]}],"x-apidog-folder":"Asist.io/Usuarios","x-apidog-status":"developing","x-run-in-apidog":"https://app.apidog.com/web/project/412546/apis/api-5849786-run","produces":["application/json"]}},"/usuario/validar/reemplazarPorToken":{"patch":{"summary":"Validar","deprecated":false,"description":"","tags":["Asist.io/Usuarios"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{},"x-apidog-ignore-properties":[],"x-apidog-orders":[]}}},"security":[],"x-apidog-folder":"Asist.io/Usuarios","x-apidog-status":"developing","x-run-in-apidog":"https://app.apidog.com/web/project/412546/apis/api-5849158-run","produces":["application/json"]}},"/usuario/olvide-mi-contrasena":{"post":{"summary":"Olvide mi contraseña","deprecated":false,"description":"","tags":["Asist.io/Usuarios"],"parameters":[{"name":"correo","in":"query","description":"","required":false,"type":"string","x-example":"enzo.meneghini@hotmail.com"},{"name":"body","in":"body","schema":{"type":"object","properties":{},"x-apidog-ignore-properties":[],"x-apidog-orders":[]}}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{},"x-apidog-ignore-properties":[],"x-apidog-orders":[]}}},"security":[],"x-apidog-folder":"Asist.io/Usuarios","x-apidog-status":"developing","x-run-in-apidog":"https://app.apidog.com/web/project/412546/apis/api-5849462-run","consumes":["application/json"],"produces":["application/json"]}},"/auth/verificar-token":{"get":{"summary":"Verificar un Token","deprecated":false,"description":"","tags":["Asist.io/Usuarios"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{},"x-apidog-ignore-properties":[],"x-apidog-orders":[]}}},"security":[{"bearer":[]}],"x-apidog-folder":"Asist.io/Usuarios","x-apidog-status":"developing","x-run-in-apidog":"https://app.apidog.com/web/project/412546/apis/api-5852155-run","produces":["application/json"]}}},"swagger":"2.0","host":"localhost:8080","schemes":["http"],"basePath":"/api/v1","definitions":{},"securityDefinitions":{"bearer":{"type":"apiKey","name":"Authorization","in":"header"}},"x-components":{}}